Created potential extensible FunctionInvokingChatClientV2
